+++ Source/ACE/Managers/WorldManager.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.Collections.Concurrent;
 using System.Diagnostics;
 using System.Linq;
 using System.Net;
@@ -9,6 +10,8 @@ using System.Threading.Tasks;
 using ACE.Common;
 using ACE.Entity;
 using ACE.Network;
+
+using ACE.Entity.Actions;
 
 using log4net;
 
@@ -31,6 +34,12 @@ namespace ACE.Managers
         public static DerethDateTime WorldStartFromTime { get; } = new DerethDateTime().UTCNowToLoreTime;
 
         public static double PortalYearTicks { get; private set; } = WorldStartFromTime.Ticks;
+
+        public static readonly ActionQueue ActionQueue = new ActionQueue();
+        public static readonly ActionQueue MotionQueue = new ActionQueue();
+        public static readonly ActionQueue BroadcastQueue = new ActionQueue();
+
+        public static readonly DelayManager DelayManager = new DelayManager();
 
         public static void Initialize()
         {
@@ -195,6 +204,12 @@ namespace ACE.Managers
 
         public static void StopWorld() { pendingWorldStop = true; }
 
+        /// <summary>
+        /// Manages updating all entities on the world.
+        ///  - Nerver-side command-line commands are handled in their own thread.
+        ///  - Network commands come from their own listener threads, and are queued in world objects
+        ///  - This thread does the rest of the work!
+        /// </summary>
         private static void UpdateWorld()
         {
             log.DebugFormat("Starting UpdateWorld thread");
@@ -205,6 +220,76 @@ namespace ACE.Managers
             {
                 worldTickTimer.Restart();
 
+                // Handle time-based timeouts
+                DelayManager.RunActions();
+
+                // Sequences of update thread:
+                // Update positions based on new tick
+                // TODO(ddevec): Physics here
+                IEnumerable<WorldObject> movedObjects = FakePhysics(PortalYearTicks);
+
+                // Do any pre-calculated landblock transfers --
+                Parallel.ForEach(movedObjects, wo =>
+                {
+                    // If it was picked up, or moved
+                    // NOTE: The object's Location can now be null, if a player logs out, or an item is picked up
+                    if (wo.Location != null && wo.Location.LandblockId != wo.CurrentLandblock.Id)
+                    {
+                        LandblockManager.RelocateObject(wo);
+                    }
+                });
+
+                // FIXME(ddevec): This O(n^2) tracking loop is a remenant of the old structure -- we should probably come up with a more efficient tracking scheme
+                Parallel.ForEach(movedObjects, mo =>
+                {
+                    // detect all world objects in ghost range
+                    List<WorldObject> woproxghost = new List<WorldObject>();
+                    woproxghost.AddRange(mo.CurrentLandblock.GetWorldObjectsInRangeForPhysics(mo, Landblock.MaxObjectGhostRange));
+
+                    // for all objects in range of this moving object or in ghost range of moving object update them.
+                    Parallel.ForEach(woproxghost, gwo =>
+                    {
+                        if (mo.Guid.IsPlayer())
+                        {
+                            // if world object is in active zone then.
+                            if (gwo.Location.SquaredDistanceTo(mo.Location) <= Landblock.MaxObjectRange * Landblock.MaxObjectRange)
+                            {
+                                // if world object is in active zone.
+                                if (!(mo as Player).GetTrackedObjectGuids().Contains(gwo.Guid))
+                                    (mo as Player).TrackObject(gwo);
+                            }
+                            // if world object is in ghost zone and outside of active zone
+                            else
+                            {
+                                if ((mo as Player).GetTrackedObjectGuids().Contains(gwo.Guid))
+                                {
+                                    (mo as Player).StopTrackingObject(gwo, false);
+                                }
+                            }
+                        }
+                    });
+                });
+
+                // Process between landblock object motions sequentially -- for safety
+                // Currently only used for picking items up off a landblock
+                MotionQueue.RunActions();
+
+                // Now, update actions within landblocks
+                //   This is responsible for updating all "actors" residing within the landblock. 
+                //   Objects and landblocks are "actors"
+                //   "actors" decide if they want to read/modify their own state (set desired velocity), move-to positions, move items, read vitals, etc
+                //   RULE: action functions CANNOT read/modify other objects 
+                //      -- unless they "own" the other object: e.g. in a container
+                //      -- action objects must send requests (actions) to other actor's action queues, and have them do modification
+                // N.B. -- Broadcasts are enqueued for sending at the end of the landblock's action time
+                // FIXME(ddevec): Goal is to eventually migrate to an "Act" function of the LandblockManager ActiveLandblocks
+                //    Inactive landblocks will be put on TimeoutManager queue for timeout killing
+                ActionQueue.RunActionsParallel();
+
+                // Handles sending out all per-landblock broadcasts -- This may rework when we rework tracking -- tbd
+                BroadcastQueue.RunActionsParallel();
+
+                // XXX(ddevec): Should this be its own step in world-update thread?
                 sessionLock.EnterReadLock();
                 try
                 {
@@ -220,6 +305,41 @@ namespace ACE.Managers
                 lastTick = (double)worldTickTimer.ElapsedTicks / Stopwatch.Frequency;
                 PortalYearTicks += lastTick;
             }
+        }
+
+        private static IEnumerable<WorldObject> FakePhysics(double timeTick)
+        {
+            ConcurrentQueue<WorldObject> movedObjects = new ConcurrentQueue<WorldObject>();
+            // Accessing ActiveLandblocks is safe here -- nothing can modify the landblocks at this point
+            Parallel.ForEach(LandblockManager.ActiveLandblocks, landblock =>
+            {
+                foreach (WorldObject wo in landblock.GetPhysicsWorldObjects())
+                {
+                    Position newPosition = wo.Location;
+                    if (wo.ForcedLocation != null)
+                    {
+                        newPosition = wo.ForcedLocation;
+                        movedObjects.Enqueue(wo);
+                    }
+                    else if (wo.RequestedLocation != null)
+                    {
+                        newPosition = wo.RequestedLocation;
+                        movedObjects.Enqueue(wo);
+                    }
+
+                    if (newPosition != wo.Location)
+                    {
+                        wo.PhysicsUpdatePosition(newPosition);
+                    }
+                }
+            });
+
+            return movedObjects;
+        }
+
+        public static double SecondsToTicks(double elapsedTimeSeconds)
+        {
+            return elapsedTimeSeconds;
         }
     }
 }