+++ Source/ACE/Network/Handlers/CharacterHandler.cs
@@ -274,14 +274,14 @@ namespace ACE.Network.Handlers
             CharacterCreateSetDefaultCharacterOptions(character);
             CharacterCreateSetDefaultCharacterPositions(character);
 
-            bool saveSuccess = await DatabaseManager.Shard.SaveObject(character);
-            // bool saveSuccess = await DbManager.SaveObject(character);
+            // We must await here -- 
+            bool saveSuccess = await DbManager.SaveObject(character);
 
-            // if (!saveSuccess)
-            // {
-            //    SendCharacterCreateResponse(session, CharacterGenerationVerificationResponse.DatabaseDown);
-            //    return;
-            // }
+            if (!saveSuccess)
+            {
+               SendCharacterCreateResponse(session, CharacterGenerationVerificationResponse.DatabaseDown);
+               return;
+            }
             // DatabaseManager.Shard.SaveCharacterOptions(character);
             // DatabaseManager.Shard.InitCharacterPositions(character);